requirements.txt
Package            Version   
------------------ ----------
black              23.12.1   
blinker            1.6.2     
click              8.1.4     
dnspython          2.4.2     
feedparser         6.0.10    
Flask              2.3.2     
gensim             4.3.2     
importlib-metadata 6.8.0     
itsdangerous       2.1.2     
Jinja2             3.1.2     
joblib             1.3.2     
MarkupSafe         2.1.3     
mypy-extensions    1.0.0     
nltk               3.8.1     
numpy              1.24.4    
packaging          23.2      
pathspec           0.12.1    
pip                20.0.2    
pkg-resources      0.0.0     
platformdirs       4.1.0     
pymongo            4.6.1     
python-dotenv      1.0.0     
regex              2023.12.25
scipy              1.10.1    
setuptools         44.0.0    
sgmllib3k          1.0.0     
smart-open         6.4.0     
tomli              2.0.1     
tqdm               4.66.1    
typing-extensions  4.9.0     
Werkzeug           2.3.6     
zipp               3.16.0    
